fix multisig
